MOD 利用openID-userId的关联方式进行查询订单

+++ src/main/java/com/jichuangtech/clothshopserver/service/UsersService.java
@@ -53,11 +53,11 @@ public class UsersService {
         return entity;
     }
 
-    public long getUserIdByOpenId(String openId) {
-        long userId = -1;
+    public int getUserIdByOpenId(String openId) {
+        int userId = -1;
         UsersEntity entity = getUserByOpenId(openId);
         if(entity != null) {
-            userId  = entity.getUserId();
+            userId  = Math.toIntExact(entity.getUserId());
         }
         LOGGER.info(" getUserIdByOpenId entity: " + entity + ", openId: " + openId);
         return userId;
